% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-load-and-processing.R
\name{extend_dataset}
\alias{extend_dataset}
\title{Extension of dataset files from folder of raw datasets}
\usage{
extend_dataset(
  input_folder,
  output_folder,
  wanted_days,
  dset_key,
  metadata_files = NULL,
  from_date = "first",
  to_date = "last",
  extend_after_end = TRUE
)
}
\arguments{
\item{input_folder}{Input folder of datasets.}

\item{output_folder}{Desired output folder of extended datasets.}

\item{wanted_days}{Minimum number of complete days of the final extended datasets.}

\item{dset_key}{Dataset key: \code{lcl} for Low Carbon London; \code{goi} for Goiener.}

\item{metadata_files}{Path or vector of paths to metadata files. Metadata files MUST contain a header and be comma-separated (",").}

\item{from_date}{Initial date and time of the interval. Either a \code{POSIXct} class in the GMT time zone OR a string \code{first}.}

\item{to_date}{Final date and time of the interval. Either a \code{POSIXct} class in the GMT time zone OR a string \code{last}.}

\item{extend_after_end}{If \code{TRUE}, the expansion is appended after the end of the time series; if \code{FALSE}, the expansion is prepended before the beginning of the time series.}
}
\value{
As many extended dataframes as files in the raw dataset folder. The dataframes are saved as R files with extension \code{.RData}.
}
\description{
Compute the extended dataframes from an input folder of raw datasets and store them in an output folder.
}
\details{
Automatizes the following sequence for a whole folder: raw dataset -> raw dataframe -> cooked dataframe -> imputed dataframe -> extended dataframe -> extended dataset.
}
