% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-load-and-processing.R
\name{extend_datasets}
\alias{extend_datasets}
\title{Extended datasets from folder of raw datasets}
\usage{
extend_datasets(input_folder, output_folder)
}
\arguments{
\item{input_folder}{Input folder of datasets.}

\item{output_folder}{Desired output folder of extended datasets.}
}
\value{
As many extended dataframes as files in the raw dataset forlder.
}
\description{
Compute the extended dataframes from an input folder of raw datasets and store them in an output folder.
}
\details{
Automatizes the following sequence for a whole folder: raw dataset -> raw dataframe -> cooked dataframe -> imputed dataframe -> extended dataframe -> extended dataset.
}
