% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-load-and-processing.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Ancillary function to extract metadata}
\usage{
extract_metadata(dfs, dset_key, filename)
}
\arguments{
\item{dset_key}{Dataset key: \code{lcl}, \code{goi}, etc.}

\item{metadata_files}{A string or a vector of strings with the file paths.}
}
\value{
A list of metadata.
}
\description{
Extract metadata from files.
}
