% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-metadata.R
\name{get_dataset_dependent_metadata}
\alias{get_dataset_dependent_metadata}
\title{Assign metadata to a list}
\usage{
get_dataset_dependent_metadata(dset_key, metadata)
}
\arguments{
\item{dset_key}{Dataset key: \code{lcl}, \code{goi}, etc.}

\item{metadata}{A dataframe of metadata.}
}
\value{
A list of metadata.
}
\description{
Extract metadata from files.
}
