% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{get_feature_names}
\alias{get_feature_names}
\title{Generation of the feature names}
\usage{
get_feature_names()
}
\value{
List of lists with the feature names.
}
\description{
Generate the names of the features.
}
\details{
NAMING FORMAT: [abs|rel]_[mean|sd|max|mean]_type_(pday)
}
