% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-tools.R
\name{get_heatmap_matrix}
\alias{get_heatmap_matrix}
\title{Get heatmap matrix}
\usage{
get_heatmap_matrix(fnames)
}
\arguments{
\item{fnames}{Two-column dataframe (data_set and file names).}
}
\value{
A matrix representing a heatmap.
}
\description{
Creation of a matrix valid to be plotted in a heatmap
}
