% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning-analysis.R
\name{pca_from_features}
\alias{pca_from_features}
\title{PCA from features}
\usage{
pca_from_features(feats_folder, ftp, otp = NULL)
}
\arguments{
\item{feats_folder}{String with the path to the folder of the features file.}

\item{ftp}{Vector of features to plot.}

\item{otp}{Vector of observations to plot. \code{NULL} indicates all observations.}
}
\value{
A list with class \code{princomp}.
}
\description{
Compute PCA from a dataframe of features.
}
