% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-tools.R
\name{plot_heatmap_matrix}
\alias{plot_heatmap_matrix}
\title{Plot heatmap matrix}
\usage{
plot_heatmap_matrix(
  m,
  format_file = NA,
  file_path = NA,
  plot_width = 800,
  plot_height = 600
)
}
\arguments{
\item{m}{Matrix to be plotted.}

\item{format_file}{Format of the output file. It can be \code{png} or \code{pdf}.}

\item{file_path}{Path to the output file INCLUDING the extension.}

\item{plot_width}{Width of the plot in pixels.}

\item{plot_height}{Heigh of the plot in pixels.}
}
\value{
Nothing, just a plot.
}
\description{
Plot a heatmap matrix representing a year of enegry consumption
}
