% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{quantiles}
\alias{quantiles}
\title{Features related to quantiles}
\usage{
quantiles(x)
}
\arguments{
\item{x}{Time series of class \code{msts}.}
}
\value{
A list with the minimum, median and maximum, quartiles, deciles, interquartile (Q3 to Q1) range and percentage of outliers.
}
\description{
Compute the minimum, median and maximum, quartiles, deciles, interquartile (Q3 to Q1) range and percentage of outliers.
}
