% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{stat_data_aggregates}
\alias{stat_data_aggregates}
\title{Features related to statistical data aggregates}
\usage{
stat_data_aggregates(x)
}
\arguments{
\item{x}{Time series of class \code{msts}.}
}
\value{
A list with statistical values of the different bins.
}
\description{
Compute means and variances of time series aggregates, where bins are days, weeks and months.
}
\details{
\code{mean_XXh} and \code{var_XXh}: the mean and variance of the time series at \code{XX} hours, where \code{XX} goes from 00 to 23. \code{unit_mean_XXh} and \code{unit_var_XXh}: the same as before but values are normalized so that the sum is 1. \code{mean_xxx} and \code{var_xxx}: the mean and variance of the time series on \code{xxx}, where \code{xxx} goes from sun (Sunday) to sat (Saturday). \code{unit_mean_xxx} and \code{unit_var_xxx}: the same as before but values are normalized so that the sum is 1. \code{mean_yyy} and \code{var_yyy}: the mean and variance of the time series in \code{yyy}, where \code{yyy} goes from jan (January) to dec (December). \code{unit_mean_yyy} and \code{unit_var_yyy}: the same as before but values are normalized so that the sum is 1.
}
