% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{stat_data_binning}
\alias{stat_data_binning}
\title{Features related to statistical data binning}
\usage{
stat_data_binning(x)
}
\arguments{
\item{x}{Time series of class `msts`.}
}
\value{
A list with statistical values of the different bins.
}
\description{
Compute means and variances of the binned signal, where bins are related to days, weeks and months
}
