% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-features.R
\name{stat_moments}
\alias{stat_moments}
\title{Features of statistical moments}
\usage{
stat_moments(x)
}
\arguments{
\item{x}{Time series of class `msts`.}
}
\value{
A list with the mean, variance, skewness and kurtosis of the time series.
}
\description{
Compute the mean, variance, skewness and kurtosis of a time series.
}
