% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-extraction.R
\name{get_extrema_dates_from_timeseries}
\alias{get_extrema_dates_from_timeseries}
\title{Initial and final dates from time series}
\usage{
get_extrema_dates_from_timeseries(tseries)
}
\arguments{
\item{tseries}{Time series of class \code{msts}.}
}
\value{
List with initial and final dates.
}
\description{
Get initial and final dates from a time series.
}
