% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-extraction.R
\name{get_features_from_cooked_dataframe}
\alias{get_features_from_cooked_dataframe}
\title{Features of a cooked dataframe}
\usage{
get_features_from_cooked_dataframe(cdf, type_of_analysis)
}
\arguments{
\item{type_of_analysis}{A string indicating the type of analysis: either \code{basic} or \code{extra}.}

\item{df}{Cooked (or extended) dataframe.}
}
\value{
List of features.
}
\description{
Get features of a cooked (or extended) dataframe.
}
