% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-extraction.R
\name{get_features_from_ext_datasets}
\alias{get_features_from_ext_datasets}
\title{Features of extended datasets in a folder}
\usage{
get_features_from_ext_datasets(
  input_folder,
  output_folder,
  type_of_analysis,
  list_of_functions = c(),
  .scale = FALSE,
  parallelize = TRUE
)
}
\arguments{
\item{input_folder}{Absolute path to the dataset folder.}

\item{output_folder}{Absolute path where the results are wanted.}

\item{type_of_analysis}{A string indicating the type of analysis: \code{basic}, \code{extra} or \code{custom}.}

\item{list_of_functions}{If \code{type_of_analysis} is \code{custom}, a list of strings indicating the functions that perform the feature extraction.}

\item{If}{\code{type_of_analysis} is \code{custom}, TRUE or FALSE indicating if the time series must be scaled to mean 0 and sd 1 prior the analysis.}
}
\value{
File containing the features
}
\description{
Get features of all extended datasets contained in a folder.
}
